Changed Event Handler to Delegate Model

+++ Warlord/Warlord/GameTools/Noise2D/PerlinNoiseSettings2D.cs
@@ -15,7 +15,7 @@ namespace GameTools.Noise2D
         public Point startingPoint;
         public double frequencyMulti;
         public double persistence;
-        public int zoom;        
+        public float zoom;        
         public int octaves;      
         public int seed;
